pyproject.toml
[project]
name = "stock-valuation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.2",
    "yfinance>=0.2.40",
    "matplotlib>=3.9.0",
    "loguru>=0.7.2",
    "click>=8.1.7",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "alpha-vantage>=3.0.0",
]

[project.scripts]
stock-valuation = "stock_valuation.__main__:_main"

[tool.uv]
package = true

[tool.ruff]
line-length = 100
lint.select = ["ALL"]
lint.ignore = [
    "ANN002", "ANN003", "ANN401", "COM812", "D203", "D205", "D213", "D402", "EXE002", "FA102",
    "ISC001", "PD901", "PGH003", "PLR0913", "S101", "TRY003"
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "ruff>=0.9.1",
]
